backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
pydantic==2.5.3
google-generativeai==0.8.6
python-dotenv==1.2.1
